requirements.txt
google-play-scraper==1.2.3
pandas>=1.3.0
python-dotenv>=0.19.0
SQLAlchemy>=1.4.0
python-dateutil>=2.8.2
# requests>=2.26.0
tqdm>=4.62.0
numpy>=1.21.0
scikit-learn>=1.0.0  # For text analysis and clustering
schedule>=1.1.0  # For scheduling weekly reports

app-store-scraper==0.3.5
textblob>=0.17.1
openpyxl>=3.0.0
google-generativeai>=0.3.0  # For LLM-powered analysis
matplotlib>=3.4.0
seaborn>=0.11.0
